backend/requirements.txt
Flask
Flask-Cors
python-dotenv
mysql-connector-python
google-generativeai # 또는 다른 LLM 라이브러리
requests